refactor: compartmentalize exception class definitions
