+++ lib/env_parser.rb
@@ -1,36 +1,10 @@
+require 'env_parser/exceptions'
 require 'env_parser/version'
 require 'active_support/all'
 
 ## The EnvParser class simplifies parsing of environment variables as different data types.
 ##
 class EnvParser
-  ## Base exception class for EnvParser.
-  ##
-  class Error < ::StandardError
-  end
-
-  ## Exception class used to indicate a type has already been defined.
-  ##
-  class TypeAlreadyDefined < Error
-  end
-
-  ## Exception class used to indicate the requested "as" type has not been defined.
-  ##
-  class UnknownType < Error
-  end
-
-  ## Exception class used to indicate value given is not convertible to the requested type.
-  ##
-  class ValueNotConvertible < Error
-  end
-
-  ## Exception class used to indicate parsed values that do not pass user-validation, either by not
-  ## being part of the given "from_set" list, or by failing the "validated_by" Proc or yield-block
-  ## check.
-  ##
-  class ValueNotAllowed < Error
-  end
-
   class << self
     ## Defines a new type for use as the "as" option on a subsequent `.parse` or `.register` call.
     ##
